server/requirements.txt
aioredis==2.0.1
annotated-types==0.7.0
anyio==4.6.0
async-timeout==4.0.3
click==8.1.7
fastapi==0.115.0
fastapi-limiter==0.1.6
fnvhash==0.1.0
h11==0.14.0
idna==3.10
pybase62==1.0.0
pydantic==2.9.2
pydantic_core==2.23.4
redis==5.1.0
sniffio==1.3.1
starlette==0.38.6
typing_extensions==4.12.2
uvicorn==0.31.0
